Bug fixes for pause-resume
- Changed `--init_pop` to accept `float` rather than `int` to remove rounding on startup
- Use `getpid()` as a temporary fix for RNG seed

+++ gillespy2/solvers/cpp/c_base/model.cpp
@@ -147,7 +147,7 @@ namespace Gillespy {
 				os << timeline[timestep] << ',';
 
 				for (int species = 0; species < model->number_species; species++) {
-					os << trajectories[trajectory][timestep][species] << ',';
+					os << (double) trajectories[trajectory][timestep][species] << ',';
 				}
 			}
 		}
